% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonction.R
\name{compter_nombre_d_adjoints}
\alias{compter_nombre_d_adjoints}
\title{Compter le nombre d'adjoints}
\usage{
compter_nombre_d_adjoints(data)
}
\arguments{
\item{data}{Un dataframe contenant au moins les colonnes \code{Nom.de.l.elu}, \code{Prenom.de.l.elu} et \code{Date.de.naissance}.}
}
\value{
Un entier representant le nombre total d'adjoints dans le dataframe.
}
\description{
Compter le nombre d'adjoints uniques
}
\details{
Cette fonction compte le nombre total d'adjoints presents dans un dataframe contenant leurs informations.
\itemize{
\item La fonction verifie la presence des colonnes obligatoires dans \code{data}.
\item Elle compte le nombre total de lignes du dataframe, supposant que chaque ligne represente un elu unique.
}
}
