% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonction.R
\name{creer_commune}
\alias{creer_commune}
\title{Cette fonction transforme un dataframe representant une seule commune en un objet de classe "commune".}
\usage{
creer_commune(df)
}
\arguments{
\item{df}{Un dataframe contenant au moins la colonne suivante :
\itemize{
\item \code{Code.de.la.commune} (character ou numeric) : Le code identifiant la commune.
}}
}
\value{
Un dataframe de classe \code{"commune"}, pret a etre utilise dans d'autres fonctions specifiques aux communes.
}
\description{
La fonction s'assure que le dataframe contient une seule commune distincte,
et lui attribue la classe \code{"commune"} s'il ne l'a pas deja.
}
\details{
\itemize{
\item La fonction verifie que la colonne \code{Code.de.la.commune} est presente dans \code{df}.
\item Elle s'assure qu'il n'y a \strong{qu'une seule commune distincte}, sinon elle genere une erreur.
\item Si la classe \code{"commune"} n'est pas deja presente, elle est ajoutee a l'objet.
}
}
