% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonction.R
\name{creer_departement}
\alias{creer_departement}
\title{Creer un Departement}
\usage{
creer_departement(df)
}
\arguments{
\item{df}{Un dataframe contenant au moins la colonne \code{Code.du.departement}.}
}
\value{
Un dataframe de classe \code{"departement"}, qui peut etre utilise avec des fonctions specifiques aux departements.
}
\description{
Cette fonction transforme un dataframe representant un departement en un objet de classe "departement".
}
\details{
\itemize{
\item La fonction verifie que la colonne \code{Code.du.departement} est presente dans \code{df}.
\item Elle s'assure que le dataframe represente un seul departement, sinon elle genere une erreur.
\item La classe \code{"departement"} est ajoutee a l'objet pour permettre un traitement specifique.
}
}
\examples{
\dontrun{
data <- data.frame(
  Code.du.departement = rep("75", 10),
  Nom.du.departement = rep("Paris", 10),
  Population = c(100, 200, 150, 180, 220, 250, 300, 270, 260, 280)
)
departement <- creer_departement(data)
class(departement)  # Verifier la classe
}

}
