% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonction.R
\name{plot_code_professions}
\alias{plot_code_professions}
\title{Visualiser la répartition des catégories socio-professionnelles des maires
Visualiser la repartition des elus par code professionnel}
\usage{
plot_code_professions(data)
}
\arguments{
\item{data}{Un dataframe contenant au moins la colonne suivante :
\itemize{
\item \code{Code.de.la.categorie.socio.professionnelle} (character) : Code professionnel des elus.
}}
}
\value{
Un graphique \code{ggplot2} representant la repartition des elus par code professionnel.
}
\description{
La fonction compte le nombre d'elus pour chaque code professionnel (\code{Code.de.la.categorie.socio.professionnelle}),
filtre les codes n'ayant aucun elu, puis produit un graphique en barres classe par ordre decroissant.
}
\details{
Cette fonction genere un graphique en barres horizontales representant le nombre d'elus par code professionnel.
\itemize{
\item Les valeurs a \code{n = 0} sont filtrees avant l'affichage du graphique.
\item Le graphique est trie par nombre d'elus de maniere decroissante.
\item Les barres sont colorees en bleu fonce pour une meilleure lisibilite.
}
}
