% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonction.R
\name{print.summary.commune}
\alias{print.summary.commune}
\title{Affichage du resume pour summary.commune}
\usage{
\method{print}{summary.commune}(x, ...)
}
\arguments{
\item{x}{Un objet de classe \code{summary.commune}.}

\item{...}{Arguments supplementaires (non utilises).}
}
\value{
NULL (resume affiche dans la console)
}
\description{
Cette fonction definit l'affichage personnalise pour les objets de classe \code{summary.commune}.
}
