% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonction.R
\name{print.summary.departement}
\alias{print.summary.departement}
\title{Affichage du resume pour summary.departement}
\usage{
\method{print}{summary.departement}(x, ...)
}
\arguments{
\item{x}{Un objet de classe \code{summary.departement}.}

\item{...}{Arguments supplementaires (non utilises).}
}
\value{
NULL (resume affiche dans la console)
}
\description{
Cette fonction definit l'affichage personnalise pour les objets de classe \code{summary.departement}.
}
