% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonction.R
\name{summary.commune}
\alias{summary.commune}
\title{Methode S3 pour summary.commune}
\usage{
\method{summary}{commune}(object, ...)
}
\arguments{
\item{object}{Un objet de classe \code{commune}.}

\item{...}{Arguments supplementaires (non utilises).}
}
\description{
Cette fonction redirige automatiquement \code{summary()} vers \code{summary_commune()}
pour les objets de classe \code{commune}.
}
