% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonction.R
\name{summary.departement}
\alias{summary.departement}
\title{Methode S3 pour summary.departement}
\usage{
\method{summary}{departement}(object, ...)
}
\arguments{
\item{object}{Un objet de classe \code{departement}.}

\item{...}{Arguments supplementaires (non utilises).}
}
\description{
Cette fonction redirige automatiquement \code{summary()} vers \code{summary_departement()}
pour les objets de classe \code{departement}.
}
