% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonction.R
\name{summary_commune}
\alias{summary_commune}
\title{Resume des donnees d'une commune}
\usage{
summary_commune(obj, ...)
}
\arguments{
\item{obj}{Un objet de classe \code{commune}.}

\item{...}{Arguments supplementaires (non utilises).}
}
\value{
Un resume des informations de la commune sous forme de liste.
}
\description{
Cette fonction genere un resume pour les objets de classe \code{commune}.
Elle affiche le nom de la commune, le nombre total d'elus et la repartition professionnelle.
}
