% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonction.R
\name{summary_departement}
\alias{summary_departement}
\title{Summary des Departements}
\usage{
summary_departement(object, ...)
}
\arguments{
\item{object}{Un objet de classe \code{departement}.}

\item{...}{Arguments supplementaires (non utilises).}
}
\value{
Un resume des informations du departement sous forme de liste.
}
\description{
Cette fonction genere un resume pour les objets de classe \code{departement}.
Elle affiche le nom du departement, le nombre total de communes, et le nombre total d'elus.
}
