% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_schema.R
\name{validate_schema}
\alias{validate_schema}
\title{Validation du schema du dataframe}
\usage{
validate_schema(df)
}
\arguments{
\item{df}{Un dataframe a valider.}
}
\value{
Aucun retour si la validation est reussie. Stoppe avec une erreur sinon.
}
\description{
Cette fonction verifie que le dataframe contient les colonnes necessaires
pour etre considere comme un objet de type "commune" ou "departement".
}
\details{
La fonction verifie la presence des colonnes essentielles :
\itemize{
\item Pour un objet "commune" : \code{Code.de.la.commune}, \code{Libelle.de.la.commune}
\item Pour un objet "departement" : \code{Code.du.departement}, \code{Libelle.du.departement}
}
}
\examples{
\dontrun{
data <- data.frame(
  Code.de.la.commune = "75056",
  Libelle.de.la.commune = "Paris"
)
validate_schema(data)
}
}
